use shutdown helper routine in main() too
At the end of main(), we wait for the other threads to join() before we shut down. As such, we're not calling exit() in shutdown. We also can't include the thread join() routines in shutdown since those threads also call shutdown. As such, I thought the best solution was just to have each thread explicitly call exit() when they want to exit, and let main handle the thread killing, etc.