Cleanup log output
The 1.12 version of p4-fusion was less verbose which had the nice property that since we always show the last log line in the clone progress UI it would contains stats on the progress.
This reverts to that, and makes the output required for our test script opt-in using a compile time flag.